requirements.txt
area==1.1.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
camelot-py==0.11.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
fonttools==4.40.0
googlemaps==4.10.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pdfminer.six==20221105
Pillow==9.5.0
psutil==5.9.5
py-console==0.1.6
pycparser==2.21
pyparsing==3.1.0
pypdf==3.10.0
PyPDF2==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
rapidfuzz==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tabulate==0.9.0
trio==0.22.0
trio-websocket==0.10.3
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.3
utils-nuuuwan==1.2.7
waitress==2.1.2
Werkzeug==2.3.6
wsproto==1.2.0
python-dotenv==1.0.0
